Make runLaunch public
